feat: Add About section to display summary from CV
This commit adds a new component called "About" to display the summary information from the CV. The component is added under the "sections" folder and is imported into the main "Section" component. The necessary changes are made to the "tsconfig.json" file to include new paths for the component imports. Additionally, new icons for Kotlin, Next.js, WorldMap, Flutter, CSS, HTML, JavaScript, Phone, and Node.js are added to the "icons" folder.

```